Refactored sec. 2.3, started para. on nothingness
Based on feedback from my essay advisor, I added an additional section
in the introduction, contextualising why contingency is the problem
that I wish to explore.

Likewise, I revisited and completely refactored the section on the
irreducibility of phenomena to consciousness. I realised it was too
specific -- and I skipped on to talking about the generation of being
from nothingness instead.