Final edits on Cosmogony, and notes for Conting.
Made a final edit to the concluding paragraph of the Cosmogony section.
It is not how I like it, and I might revisit it later -- but I will
definitely move on now, to the Contingency of Human existence. I added
some notes for that section based on my talk with my advisor. Now all
that remains is to begin it.