Beginning of major rewrite on Chpt. 2: Contingn.
This commit demarcates the start of a major rewrite of the chapter on
contingency. After meeting with my advisor, I have decided to change
the direction of my inquiry starting at Chapter 2 to focus more on the
possibility of love and philosophy under Sartrean ontology.

This means I'll have to significantly redo Chapter 2, so I have
copied the old .tex files to contingency-old.tex, and recreated a
blank document for contingency.tex.

I will be making modifications on to contingency.tex in the direction
of the rewrite following this commit.